Relaciones entre tablas

+++ project/backend/src/main/java/com/jpa/backend/repositories/CentroRepository.java
@@ -3,6 +3,6 @@ package com.jpa.backend.repositories;
 import com.jpa.backend.entities.Centro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CentroRepository extends JpaRepository<Centro, Integer> {
+public interface CentroRepository extends JpaRepository<Centro, Long> {
 
 }
