+++ project/backend/src/main/java/com/jpa/backend/repositories/DestinatarioRepository.java
@@ -3,5 +3,5 @@ package com.jpa.backend.repositories;
 import com.jpa.backend.entities.Destinatario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DestinatarioRepository extends JpaRepository<Destinatario, Integer> {
+public interface DestinatarioRepository extends JpaRepository<Destinatario, Long> {
 }
