+++ project/backend/src/main/java/com/jpa/backend/repositories/GerenteRepository.java
@@ -3,5 +3,5 @@ package com.jpa.backend.repositories;
 import com.jpa.backend.entities.Gerente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GerenteRepository extends JpaRepository<Gerente, Integer> {
+public interface GerenteRepository extends JpaRepository<Gerente, Long> {
 }
